refactor: migrate google sheet commands
