+++ src/extensions/google-sheet-logging/google-sheet-constants/google-sheet-interaction-names.ts
@@ -1,0 +1,6 @@
+enum GoogleSheetCommands {
+    stats = 'stats',
+    weekly_report = 'weekly_report'
+}
+
+export { GoogleSheetCommands };
